Add DeveloperToolVoter
