+++ Backoffice/Security/Authorization/Voter/AbstractPerimeterVoter.php
@@ -2,7 +2,6 @@
 
 namespace OpenOrchestra\Backoffice\Security\Authorization\Voter;
 
-use OpenOrchestra\Backoffice\Security\ContributionActionInterface;
 use OpenOrchestra\Backoffice\Model\PerimeterInterface;
 
 /**
@@ -12,47 +11,6 @@ use OpenOrchestra\Backoffice\Model\PerimeterInterface;
  */
 abstract class AbstractPerimeterVoter extends AbstractVoter
 {
-    /**
-     * Return the list of supported classes
-     *
-     * @return array
-     */
-    abstract protected function getSupportedClasses();
-
-    /**
-     * @return array
-     */
-    protected function getSupportedAttributes()
-    {
-        return array(
-            ContributionActionInterface::READ,
-            ContributionActionInterface::ADD,
-            ContributionActionInterface::EDIT,
-            ContributionActionInterface::DELETE
-        );
-    }
-
-    /**
-     * @param string $attribute
-     * @param mixed  $subject
-     *
-     * return bool
-     */
-    protected function supports($attribute, $subject)
-    {
-        if (!in_array($attribute, $this->getSupportedAttributes())) {
-            return false;
-        }
-
-        foreach ($this->getSupportedClasses() as $supportedClass) {
-            if ($subject instanceof $supportedClass) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
     /**
      * Check if $subjectKey is in an allowed perimeter to $user
      * The perimeter to check is of $entityType
