+++ Backoffice/Security/Authorization/Voter/AbstractVoter.php
@@ -4,12 +4,54 @@ namespace OpenOrchestra\Backoffice\Security\Authorization\Voter;
 
 use Symfony\Component\Security\Core\Authorization\Voter\Voter;
 use OpenOrchestra\Backoffice\Security\ContributionRoleInterface;
+use OpenOrchestra\Backoffice\Security\ContributionActionInterface;
 
 /**
  * Class AbstractVoter
  */
 abstract class AbstractVoter extends Voter
 {
+    /**
+     * Return the list of supported classes
+     *
+     * @return array
+     */
+    abstract protected function getSupportedClasses();
+
+    /**
+     * @return array
+     */
+    protected function getSupportedAttributes()
+    {
+        return array(
+            ContributionActionInterface::READ,
+            ContributionActionInterface::ADD,
+            ContributionActionInterface::EDIT,
+            ContributionActionInterface::DELETE
+        );
+    }
+
+    /**
+     * @param string $attribute
+     * @param mixed  $subject
+     *
+     * return bool
+     */
+    protected function supports($attribute, $subject)
+    {
+        if (!in_array($attribute, $this->getSupportedAttributes())) {
+            return false;
+        }
+
+        foreach ($this->getSupportedClasses() as $supportedClass) {
+            if ($subject instanceof $supportedClass) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
     /**
      * @param UserInterface|string $user
      *
