+++ BackofficeBundle/Resources/config/voter.yml
@@ -1,8 +1,9 @@
 parameters:
-    open_orchestra_backoffice.voter.node.class:          OpenOrchestra\Backoffice\Security\Authorization\Voter\NodeVoter
-    open_orchestra_backoffice.voter.content.class:       OpenOrchestra\Backoffice\Security\Authorization\Voter\ContentVoter
-    open_orchestra_backoffice.voter.trash_item.class:    OpenOrchestra\Backoffice\Security\Authorization\Voter\TrashItemVoter
-    open_orchestra_backoffice.voter.configuration.class: OpenOrchestra\Backoffice\Security\Authorization\Voter\ConfigurationVoter
+    open_orchestra_backoffice.voter.node.class:           OpenOrchestra\Backoffice\Security\Authorization\Voter\NodeVoter
+    open_orchestra_backoffice.voter.content.class:        OpenOrchestra\Backoffice\Security\Authorization\Voter\ContentVoter
+    open_orchestra_backoffice.voter.trash_item.class:     OpenOrchestra\Backoffice\Security\Authorization\Voter\TrashItemVoter
+    open_orchestra_backoffice.voter.configuration.class:  OpenOrchestra\Backoffice\Security\Authorization\Voter\ConfigurationVoter
+    open_orchestra_backoffice.voter.developer_tool.class: OpenOrchestra\Backoffice\Security\Authorization\Voter\DeveloperToolVoter
 
 services:
     open_orchestra_backoffice.voter.node:
@@ -19,5 +20,9 @@ services:
             - { name: security.voter }
     open_orchestra_backoffice.voter.configuration:
        class: '%open_orchestra_backoffice.voter.configuration.class%'
+       tags:
+            - { name: security.voter }
+    open_orchestra_backoffice.voter.developer_tool:
+       class: '%open_orchestra_backoffice.voter.developer_tool.class%'
        tags:
             - { name: security.voter }
